refactor: reuse function to create mocked client
